training: cast labels to long for BinaryAccuracy under bf16 data
The bf16-mixed default makes the data loader emit bfloat16 tensors,
including labels.  BinaryAccuracy.update() validates its target with
torch.unique, and torch.unique is not implemented for bfloat16:

    RuntimeError: "unique" not implemented for 'BFloat16'

All the other metric calls in _log already cast labels explicitly
(.int() / .long() / .float()); only the self.acc call was missing
the cast, so the issue had been latent until the data dtype changed.
Add labels.long() to both the train/val and the test branches to
bring it in line with the rest.

+++ varformer/training/lightning_module.py
@@ -209,7 +209,7 @@ class VarformerLightningModule(pl.LightningModule):
                 labels = labels.unsqueeze(0)
         if step_type in ["train", "val"]:
             self.log(f"{step_type}_loss", loss, batch_size=labels.shape[0])
-            self.log(f"{step_type}_acc", self.acc(bin_preds, labels), batch_size=labels.shape[0])
+            self.log(f"{step_type}_acc", self.acc(bin_preds, labels.long()), batch_size=labels.shape[0])
             self.log(f"{step_type}_auroc", self.auroc(probas, labels.int()), batch_size=labels.shape[0])
             spearman_val = self.spearman(probas, labels.float())
             self.log(f"{step_type}_spearman", spearman_val, batch_size=labels.shape[0])
@@ -218,7 +218,7 @@ class VarformerLightningModule(pl.LightningModule):
             self.log(f"{step_type}_f1", self.f1(bin_preds, labels.long()), batch_size=labels.shape[0])
             self.log(f"{step_type}_auprc", self.auprc(probas, labels.long()), batch_size=labels.shape[0])
         else:
-            self.log(f"{step_type}_acc_{test_source}", self.acc(bin_preds, labels), batch_size=labels.shape[0])
+            self.log(f"{step_type}_acc_{test_source}", self.acc(bin_preds, labels.long()), batch_size=labels.shape[0])
             self.log(f"{step_type}_auroc_{test_source}", self.auroc(bin_preds, labels.int()), batch_size=labels.shape[0])
             self.log(f"{step_type}_spearman_{test_source}", self.spearman(probas, labels.float()), batch_size=labels.shape[0])
             self.log(f"{step_type}_recall_{test_source}", self.recall(bin_preds, labels.long()), batch_size=labels.shape[0])
